rename hadoop integration tests
