+++ integration-tests/src/test/java/tachyon/hadoop/contract/TachyonFSContractCreateIntegrationTest.java
@@ -16,10 +16,10 @@
 package tachyon.hadoop.contract;
 
 import org.apache.hadoop.conf.Configuration;
-import org.apache.hadoop.fs.contract.AbstractContractRenameTest;
+import org.apache.hadoop.fs.contract.AbstractContractCreateTest;
 import org.apache.hadoop.fs.contract.AbstractFSContract;
 
-public class TestTachyonFSContractRename extends AbstractContractRenameTest {
+public class TachyonFSContractCreateIntegrationTest extends AbstractContractCreateTest {
   @Override
   protected AbstractFSContract createContract(Configuration conf) {
     return new TachyonFSContract(conf);
